# swagger.json
{
    "swagger": "2.0",
    "info": {
      "version": "1.0.0",
      "title": "Jason's Calculator Project",
      "description": "My User Project Application API",
      "license": {
        "name": "MIT",
        "url": "https://opensource.org/licenses/MIT"
      }
    },
    "host": "localhost:3000",
    "basePath": "/",
    "tags": [
      {
        "name": "Calculator",
        "description": "API for Calculus in the system"
      }
    ],
    "schemes": ["http"],
    "consumes": ["application/json"],
    "produces": ["application/json"],
    "paths": {
      "/calculator/add": {
        "get": {
          "tags": ["Calculator"],
          "summary": "Add 2 numbers",
          "parameters": [
            {
              "name": "num1",
              "in": "query",
              "description": "The First Number"
            },
            {
              "name": "num2",
              "in": "query",
              "description": "The Second Number"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with the sum of the two numbers"
            },
            "default": {
              "description": "Unexpected error"
            }
          }
        }
      },
      "/calculator/subtract": {
        "get": {
          "tags": ["Calculator"],
          "summary": "Subtract second number from first number",
          "parameters": [
            {
              "name": "num1",
              "in": "query",
              "description": "The First Number"
            },
            {
              "name": "num2",
              "in": "query",
              "description": "The Second Number"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with the difference of the two numbers"
            },
            "default": {
              "description": "Unexpected error"
            }
          }
        }
      },
      "/calculator/multiply": {
        "get": {
          "tags": ["Calculator"],
          "summary": "Multiply 2 numbers",
          "parameters": [
            {
              "name": "num1",
              "in": "query",
              "description": "The First Number"
            },
            {
              "name": "num2",
              "in": "query",
              "description": "The Second Number"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with the product of the two numbers"
            },
            "default": {
              "description": "Unexpected error"
            }
          }
        }
      },
      "/calculator/divide": {
        "get": {
          "tags": ["Calculator"],
          "summary": "Divide first number by second number",
          "parameters": [
            {
              "name": "num1",
              "in": "query",
              "description": "The numerator"
            },
            {
              "name": "num2",
              "in": "query",
              "description": "The denominator"
            }
          ],
          "responses": {
            "200": {
              "description": "Successful response with the quotient of the two numbers"
            },
            "400": {
              "description": "Invalid input, cannot divide by zero"
            },
            "default": {
              "description": "Unexpected error"
            }
          }
        }
      }
    }
  }